LPS-48425 Update ServiceBuilder to remove unused files if possible - this change gets rid of LocalService, Service, Finder, and ActionableDynamicQuery files generated by service builder.

+++ portal-impl/src/com/liferay/portal/tools/servicebuilder/ServiceBuilder.java
@@ -734,6 +734,13 @@ public class ServiceBuilder {
 								_createExportActionableDynamicQuery(entity);
 							}
 						}
+						else {
+							_removeActionableDynamicQuery(entity);
+
+							if (entity.isStagedModel()) {
+								_removeExportActionableDynamicQuery(entity);
+							}
+						}
 
 						if (entity.hasColumns()) {
 							_createHbm(entity);
@@ -788,6 +795,17 @@ public class ServiceBuilder {
 								entity, _SESSION_TYPE_LOCAL);
 							_createServiceWrapper(entity, _SESSION_TYPE_LOCAL);
 						}
+						else {
+							_removeServiceImpl(entity, _SESSION_TYPE_LOCAL);
+							_removeServiceBaseImpl(entity, _SESSION_TYPE_LOCAL);
+							_removeService(entity, _SESSION_TYPE_LOCAL);
+							_removeServiceUtil(entity, _SESSION_TYPE_LOCAL);
+
+							_removeServiceClp(entity, _SESSION_TYPE_LOCAL);
+							_removeServiceClpInvoker(
+								entity, _SESSION_TYPE_LOCAL);
+							_removeServiceWrapper(entity, _SESSION_TYPE_LOCAL);
+						}
 
 						if (entity.hasRemoteService()) {
 							_createServiceImpl(entity, _SESSION_TYPE_REMOTE);
@@ -813,6 +831,23 @@ public class ServiceBuilder {
 							}
 
 							_createServiceSoap(entity);
+						}
+						else {
+							_removeServiceImpl(entity, _SESSION_TYPE_REMOTE);
+							_removeServiceBaseImpl(entity, _SESSION_TYPE_REMOTE);
+							_removeService(entity, _SESSION_TYPE_REMOTE);
+							_removeServiceUtil(entity, _SESSION_TYPE_REMOTE);
+
+							_removeServiceClp(entity, _SESSION_TYPE_REMOTE);
+							_removeServiceClpInvoker(
+								entity, _SESSION_TYPE_REMOTE);
+							_removeServiceWrapper(entity, _SESSION_TYPE_REMOTE);
+
+							if (Validator.isNotNull(_remotingFileName)) {
+								_removeServiceHttp(entity);
+							}
+
+							_removeServiceSoap(entity);
 						}
 					}
 					else {
@@ -2113,6 +2148,10 @@ public class ServiceBuilder {
 
 	private void _createFinder(Entity entity) throws Exception {
 		if (!entity.hasFinderClass()) {
+			//If no Finder is needed, remove any previously existing versions
+
+			_removeFinder(entity);
+
 			return;
 		}
 
@@ -2152,6 +2191,10 @@ public class ServiceBuilder {
 
 	private void _createFinderUtil(Entity entity) throws Exception {
 		if (!entity.hasFinderClass()) {
+			//If no Finder is needed, remove any previously existing versions
+
+			_removeFinderUtil(entity);
+
 			return;
 		}
 
@@ -3215,6 +3258,28 @@ public class ServiceBuilder {
 			if (ejbFile.exists()) {
 				System.out.println("Relocating " + ejbFile);
 
+				ejbFile.delete();
+			}
+		}
+	}
+
+private void _removeServiceUtil(Entity entity, int sessionType)
+		throws Exception {
+
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/" + entity.getName() +
+				_getSessionTypeName(sessionType) + "ServiceUtil.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+
+		if (!_serviceOutputPath.equals(_outputPath)) {
+			ejbFile = new File(
+				_outputPath + "/service/" + entity.getName() +
+					_getSessionTypeName(sessionType) + "ServiceUtil.java");
+
+			if (ejbFile.exists()) {
 				ejbFile.delete();
 			}
 		}
@@ -5104,6 +5169,179 @@ public class ServiceBuilder {
 		StringUtil.readLines(classLoader.getResourceAsStream(fileName), lines);
 
 		return lines;
+	}
+
+	private void _removeActionableDynamicQuery(Entity entity) throws Exception {
+		if (_osgiModule) {
+			return;
+		}
+
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/persistence/" +
+				entity.getName() + "ActionableDynamicQuery.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeExportActionableDynamicQuery(Entity entity)
+		throws Exception {
+
+		if (_osgiModule) {
+			return;
+		}
+
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/persistence/" +
+				entity.getName() + "ExportActionableDynamicQuery.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeFinder(Entity entity) throws Exception {
+
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/persistence/" + entity.getName() +
+				"Finder.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+
+		if (!_serviceOutputPath.equals(_outputPath)) {
+			ejbFile = new File(
+				_outputPath + "/service/persistence/" + entity.getName() +
+					"Finder.java");
+
+			if (ejbFile.exists()) {
+				ejbFile.delete();
+			}
+		}
+	}
+
+	private void _removeFinderUtil(Entity entity) throws Exception {
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/persistence/" + entity.getName() +
+				"FinderUtil.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+
+		if (!_serviceOutputPath.equals(_outputPath)) {
+			ejbFile = new File(
+				_outputPath + "/service/persistence/" + entity.getName() +
+					"FinderUtil.java");
+
+			if (ejbFile.exists()) {
+				ejbFile.delete();
+			}
+		}
+	}
+
+	private void _removeService(Entity entity, int sessionType)
+		throws Exception {
+
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/" + entity.getName() +
+				_getSessionTypeName(sessionType) + "Service.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+
+		if (!_serviceOutputPath.equals(_outputPath)) {
+			ejbFile = new File(
+				_outputPath + "/service/" + entity.getName() +
+					_getSessionTypeName(sessionType) + "Service.java");
+
+			if (ejbFile.exists()) {
+				ejbFile.delete();
+			}
+		}
+	}
+
+	private void _removeServiceBaseImpl(Entity entity, int sessionType)
+		throws Exception {
+
+		File ejbFile = new File(
+			_outputPath + "/service/base/" + entity.getName() +
+				_getSessionTypeName(sessionType) + "ServiceBaseImpl.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeServiceClp(Entity entity, int sessionType)
+		throws Exception {
+
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/" + entity.getName() +
+				_getSessionTypeName(sessionType) + "ServiceClp.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeServiceClpInvoker(Entity entity, int sessionType)
+		throws Exception {
+
+		File ejbFile = new File(
+			_outputPath + "/service/base/" + entity.getName() +
+				_getSessionTypeName(sessionType) + "ServiceClpInvoker.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeServiceHttp(Entity entity) throws Exception {
+		File ejbFile = new File(
+			_outputPath + "/service/http/" + entity.getName() +
+				"ServiceHttp.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeServiceImpl(Entity entity, int sessionType)
+		throws Exception {
+
+		File ejbFile = new File(
+			_outputPath + "/service/impl/" + entity.getName() +
+				_getSessionTypeName(sessionType) + "ServiceImpl.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeServiceSoap(Entity entity) throws Exception {
+		File ejbFile = new File(
+			_outputPath + "/service/http/" + entity.getName() +
+				"ServiceSoap.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
+	}
+
+	private void _removeServiceWrapper(Entity entity, int sessionType)
+		throws Exception {
+
+		File ejbFile = new File(
+			_serviceOutputPath + "/service/" + entity.getName() +
+				_getSessionTypeName(sessionType) + "ServiceWrapper.java");
+
+		if (ejbFile.exists()) {
+			ejbFile.delete();
+		}
 	}
 
 	private void _resolveEntity(Entity entity) throws IOException {
